Move context utils to separate package
