+++ pkg/kubeconfig/context_test.go
@@ -1,4 +1,4 @@
-package login
+package kubeconfig
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 
 func Test_generateKubeContextName(t *testing.T) {
 	codeName := "test"
-	result := generateKubeContextName(codeName)
+	result := GenerateKubeContextName(codeName)
 	expected := "gs-test"
 
 	if result != expected {
@@ -39,7 +39,7 @@ func Test_isKubeContext(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			result := isKubeContext(tc.input)
+			result := IsKubeContext(tc.input)
 
 			if result != tc.expected {
 				t.Fatalf("Value not expected, got: %t", result)
@@ -73,7 +73,7 @@ func Test_getCodeNameFromKubeContext(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			result := getCodeNameFromKubeContext(tc.input)
+			result := GetCodeNameFromKubeContext(tc.input)
 
 			if result != tc.expected {
 				t.Fatalf("Value not expected, got: %s", result)
@@ -112,7 +112,7 @@ func Test_isCodeName(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			result := isCodeName(tc.input)
+			result := IsCodeName(tc.input)
 
 			if result != tc.expected {
 				t.Fatalf("Value not expected, got: %t", result)
